Minor refactoring after PR review
