+++ fork-client/build.gradle
@@ -1,10 +1,5 @@
 apply plugin: 'java-library'
 apply from: '../build-scripts/gradle-mvn-push.gradle'
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-}
 
-//noinspection GroovyUnusedAssignment
-sourceCompatibility = rootProject.sourceCompatibility
-//noinspection GroovyUnusedAssignment
-targetCompatibility = rootProject.sourceCompatibility
+sourceCompatibility = JavaVersion.VERSION_1_7
+targetCompatibility = JavaVersion.VERSION_1_7
