+++ fork-runner/src/main/java/com/shazam/fork/runner/TestRun.java
@@ -13,6 +13,7 @@ import com.android.ddmlib.*;
 import com.android.ddmlib.testrunner.*;
 import com.google.common.base.Strings;
 import com.shazam.fork.model.TestCaseEvent;
+import com.shazam.fork.system.PermissionGrantingManager;
 import com.shazam.fork.system.io.RemoteFileManager;
 
 import org.slf4j.Logger;
@@ -21,8 +22,6 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.List;
 
-import static com.shazam.fork.PermissionUtils.grantAllPermissionsIfAllowed;
-import static com.shazam.fork.PermissionUtils.revokePermissions;
 import static com.shazam.fork.injector.ConfigurationInjector.configuration;
 import static java.lang.String.format;
 
@@ -31,13 +30,16 @@ class TestRun {
     private final String poolName;
 	private final TestRunParameters testRunParameters;
 	private final List<ITestRunListener> testRunListeners;
+	private final PermissionGrantingManager permissionGrantingManager;
 
-    public TestRun(String poolName,
+	public TestRun(String poolName,
 				   TestRunParameters testRunParameters,
-				   List<ITestRunListener> testRunListeners) {
+				   List<ITestRunListener> testRunListeners,
+				   PermissionGrantingManager permissionGrantingManager) {
         this.poolName = poolName;
 		this.testRunParameters = testRunParameters;
 		this.testRunListeners = testRunListeners;
+		this.permissionGrantingManager = permissionGrantingManager;
 	}
 
 	public void execute() {
@@ -70,16 +72,17 @@ class TestRun {
 		}
 
 		List<String> permissionsToRevoke = testRunParameters.getTest().getPermissionsToRevoke();
-		revokePermissions(testRunParameters.getApplicationPackage(), testRunParameters.getDeviceInterface(), permissionsToRevoke);
+
+		permissionGrantingManager.revokePermissions(testRunParameters.getApplicationPackage(), testRunParameters.getDeviceInterface(), permissionsToRevoke);
 
 		try {
 			runner.run(testRunListeners);
 		} catch (ShellCommandUnresponsiveException | TimeoutException e) {
 			logger.warn("Test: " + testClassName + " got stuck. You can increase the timeout in settings if it's too strict");
 		} catch (AdbCommandRejectedException | IOException e) {
-		throw new RuntimeException(format("Error while running test %s %s", test.getTestClass(), test.getTestMethod()), e);
+			throw new RuntimeException(format("Error while running test %s %s", test.getTestClass(), test.getTestMethod()), e);
 		} finally {
-			grantAllPermissionsIfAllowed(configuration(), testRunParameters.getApplicationPackage(), testRunParameters.getDeviceInterface());
+			permissionGrantingManager.grantAllPermissionsIfAllowed(configuration(), testRunParameters.getApplicationPackage(), testRunParameters.getDeviceInterface());
 		}
 
     }
