+++ fork-runner/src/main/java/com/shazam/fork/runner/TestRunFactory.java
@@ -14,11 +14,13 @@ import com.android.ddmlib.testrunner.ITestRunListener;
 import com.shazam.fork.Configuration;
 import com.shazam.fork.model.*;
 import com.shazam.fork.runner.listeners.TestRunListenersFactory;
+import com.shazam.fork.system.PermissionGrantingManager;
 
 import java.util.List;
 import java.util.Queue;
 
 import static com.shazam.fork.runner.TestRunParameters.Builder.testRunParameters;
+import static com.shazam.fork.system.PermissionGrantingManager.permissionGrantingManager;
 
 public class TestRunFactory {
 
@@ -57,6 +59,7 @@ public class TestRunFactory {
         return new TestRun(
                 pool.getName(),
                 testRunParameters,
-                testRunListeners);
+                testRunListeners,
+                permissionGrantingManager());
     }
 }
